[diffusion] extract overlay materialization helpers
